Moved TestBase to Component subdirectory
